refactor: rename pack path
